Residual pass: three more cross-variant/wrapper visibility gaps in runtime -tests
Continuing the wave-plan-ruled 29-error pre-existing-defect pass. Three
distinct, independently-verified fixes, taking runtime's -tests build from
31 errors down to 21:

1. A named untyped constant written with Go's exponent float-literal syntax
   (`const gcCPULimiterUpdatePeriod = 10e6`) is UntypedFloat-kinded even
   though its value is a whole number -- go/types classifies by literal
   syntax, not value. A bare/computed reference to it therefore renders as
   the golib UntypedFloat wrapper struct, which had no bridge to any
   integer-based defined type (Duration and siblings) -- only UntypedInt
   did, per the generator's own now-stale "floats keep their existing
   routes" comment. Added the UntypedFloat twin of that bridge operator,
   safe by construction: go/types only ever accepts the conversion at an
   integer-typed call site when the constant's value already IS exactly
   representable as one, so the generated bridge can never truncate a
   fraction Go itself would have rejected.

2. A comparison whose one side is concrete (`l.Capacity()`, ulong) and
   whose other side is a computed expression mixing TWO DIFFERENT untyped
   wrapper kinds (`procs * capacityPerProc`, UntypedInt * UntypedFloat) has
   neither an operator of its own nor a concrete operand for the existing
   arithmetic-side cast logic to key off, so it silently resolves through
   whichever primitive both wrappers happen to share first (int32) --
   comparing that against the genuinely concrete ulong is then CS0034,
   ambiguous between widening up and narrowing down. Comparison operators
   now get the identical treatment arithmetic operators already had for a
   COMPUTED untyped-const operand; a bare named-const reference (which
   resolves fine on its own) is unaffected.

3. A type alias a `_test.go` file declares (`type G = g`, export_test.go)
   is emitted as its own `global using` -- compilation-scoped, member of no
   class -- but two independent renderers were bridge-qualifying it anyway:
   typeNameResolution.go's types.Type-based path already had the bare-alias
   rule (testDeclaredAliasSpelledBare) for cross-package references, but
   convIdent.go's sibling path (whiteboxBridgeMember, for a bare/dot-
   imported identifier) had no equivalent check at all and always
   qualified. Render it bare there too, matching the internal variant's own
   spelling of the same alias.

While investigating (2), found and fixed the same-shaped gap one layer up:
a package-level var/const can ALSO collide with an unrelated method
(`export_test.go`'s `var Lock = lock` collides with RWMutex's own `Lock`
method, Δ-renaming the var) -- the identical cross-variant visibility bug
Wave A1 fixed for types, just never extended to non-type declarations.
testTypeRenames (session-scoped, object-keyed) now also registers a
renamed var/const at its own declaration site; whiteboxBridgeMember
consults it for the *types.Var/*types.Const case the same way it already
did for *types.TypeName. This did NOT resolve metrics_test.cs's own Lock
call site (a different, not-yet-found declaring code path also computes
the bridge-qualified reference) -- kept as a real, still-correct
improvement for whatever construct DOES reach visitValueSpec's ordinary
package-level loop, not reverted, but not claimed as metrics_test.cs's fix.

Two behavioral-test goldens updated (UntypedFloatConstExpr, IpAdapterAddresses)
-- both changes verified behaviorally correct first (Output phase: C# vs
`go run` stdout, 0 failures) before accepting the Target mismatch as an
intended golden update rather than a regression.

Remaining 21 errors regroup into roughly six root causes, not 21 independent
ones -- notably a [GoRecv]/method-expression gap (Go's `(*T).Method` on a
value-receiver method needing a pointer-receiver overload the generator
never produces for a whitebox test type) accounts for 9 of them across
stack_test.cs and traceback_test.cs, and a separately-lifted-anonymous-
type identity gap accounts for 3 more across hash_test.cs and
traceback_test.cs. Neither touched here.

Gates: converter go test ./... + go vet clean (109s). Full go2cs-stdlib.slnx
build: 307/307, 0 errors (verifies the UntypedFloat bridge operator touches
nothing else in the corpus). CNR: 2 intended golden changes, reproduced
identically across two runs, both independently verified behaviorally
correct via Output-phase comparison before the goldens were updated.

+++ src/tests/Behavioral/IpAdapterAddresses/main.cs
@@ -175,10 +175,10 @@ internal static void Main() {
                 masksValid = false;
                 continue;
             }
-            if ((~a).IP.To4() != default! && bits != 8 * Δnet.IPv4len) {
+            if ((~a).IP.To4() != default! && bits != (nint)(8 * Δnet.IPv4len)) {
                 masksValid = false;
             }
-            if ((~a).IP.To4() == default! && bits != 8 * Δnet.IPv6len) {
+            if ((~a).IP.To4() == default! && bits != (nint)(8 * Δnet.IPv6len)) {
                 masksValid = false;
             }
             if ((~a).IP.IsLoopback()){
